fix type mismatch

+++ backends/sve/ceed-sve-tensor-f64.c
@@ -40,7 +40,7 @@ static inline int CeedTensorContract_Sve_Blocked(CeedTensorContract contract,
       for (CeedInt j=0; j<(J/JJ)*JJ; j+=JJ)
         for (CeedInt jj=0; jj<JJ; jj++) { // unroll
           // C vectorization by compiler
-          int64_t i = (a*J+j+jj)*C;
+          int32_t i = (a*J+j+jj)*C;
           svbool_t pg = svwhilelt_b64(i, C);
           do {
             // Load u, v into vectors
@@ -63,7 +63,7 @@ static inline int CeedTensorContract_Sve_Blocked(CeedTensorContract contract,
         // Blocks of JJ rows
           for (CeedInt jj=0; jj<J-j; jj++) { // not unrolled
             // C vectorization by compiler
-            int64_t i = (a*J+j+jj)*C;
+            int32_t i = (a*J+j+jj)*C;
             svbool_t pg = svwhilelt_b64(i, C);
             do {
               // Load u, v into vectors
